msm: kgsl: check if state change to suspend succeeds
Check the return value of kgsl_pwrctrl_change_state()
when requested state is suspend. This call can fail
and we encounter adverse effects, if we proceed without
checking if the call succeeded. For instance, in case of
_preemption_store(), it is fatal to change ringbuffer,
when state change to suspend fails. Also add a helper
function that changes a given flag safely, only after
successful state change to suspend.

Change-Id: I8a929cc201d127baeaabed194655ac73a8032b2f
Signed-off-by: Puranam V G Tejaswi <pvgtejas@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>